requirements.txt
anyio==4.4.0
asgiref==3.8.1
attrs==24.2.0
black==24.8.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Django==5.1
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
PyJWT==2.9.0
pyTelegramBotAPI==4.22.1
python-dotenv==1.0.1
python-telegram-bot==21.4
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
sniffio==1.3.1
sqlparse==0.5.1
stripe==10.8.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
